+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/service/TaskMaintainService.java
@@ -22,4 +22,23 @@ public interface TaskMaintainService extends BatchCrudService<LongIdKey, Task>,
 
     String CREATE_DATE_DESC = "create_date_desc";
     String CHILD_FOR_SECTION_CREATE_DATE_DESC = "child_for_section_create_date_desc";
+
+    /**
+     * 获取将要被清理的任务实体。
+     *
+     * <p>
+     * 返回 <code>endedDate（结束日期）</code> 早于指定的日期的实体，
+     * 且 <code>status（状态）</code> 在 <code>任务完成, 任务失败, 任务过期, 任务死亡</code> 中，
+     * 且按照 <code>endedDate（结束日期）</code> 升序排列。
+     *
+     * <p>
+     * 参数列表：
+     * <ol>
+     *     <li>Data 指定的日期。</li>
+     * </ol>
+     * 返回的数据按照 <code>endedDate（结束日期）</code> 升序排列。
+     *
+     * @since 2.2.0
+     */
+    String TO_PURGED = "to_purged";
 }